function-calling-app/requirements.txt
# OpenAI Function Calling Weather Assistant Dependencies

# Core Dependencies
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0